backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.21.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
celery==5.3.4
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
supabase==2.3.5
cryptography==41.0.7
websocket-client==1.7.0
httpx==0.25.2
python-multipart==0.0.6
